Move ./pkg/action to ./action
