+++ action/chatgpt/testdata/aws_guardduty_example.json
