+++ action/jira/add_attachment_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/m-mizutani/gt"
-	"github.com/secmon-lab/alertchain/pkg/action/jira"
+	"github.com/secmon-lab/alertchain/action/jira"
 	"github.com/secmon-lab/alertchain/pkg/domain/model"
 	"github.com/secmon-lab/alertchain/pkg/utils"
 )
