+++ action/jira/add_comment_test.go
@@ -7,7 +7,7 @@ import (
 	go_jira "github.com/andygrunwald/go-jira"
 
 	"github.com/m-mizutani/gt"
-	"github.com/secmon-lab/alertchain/pkg/action/jira"
+	"github.com/secmon-lab/alertchain/action/jira"
 	"github.com/secmon-lab/alertchain/pkg/domain/model"
 	"github.com/secmon-lab/alertchain/pkg/utils"
 )
