+++ action/jira/create_issue_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/m-mizutani/gt"
-	"github.com/secmon-lab/alertchain/pkg/action/jira"
+	"github.com/secmon-lab/alertchain/action/jira"
 	"github.com/secmon-lab/alertchain/pkg/domain/model"
 	"github.com/secmon-lab/alertchain/pkg/utils"
 )
