+++ action/opsgenie/create_alert_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/m-mizutani/gt"
 	"github.com/opsgenie/opsgenie-go-sdk-v2/alert"
-	"github.com/secmon-lab/alertchain/pkg/action/opsgenie"
+	"github.com/secmon-lab/alertchain/action/opsgenie"
 	"github.com/secmon-lab/alertchain/pkg/domain/model"
 	"github.com/secmon-lab/alertchain/pkg/utils"
 )
