+++ action/otx/action_test.go
@@ -9,7 +9,7 @@ import (
 	"testing"
 
 	"github.com/m-mizutani/gt"
-	"github.com/secmon-lab/alertchain/pkg/action/otx"
+	"github.com/secmon-lab/alertchain/action/otx"
 	"github.com/secmon-lab/alertchain/pkg/domain/model"
 	"github.com/secmon-lab/alertchain/pkg/domain/types"
 )
